some of the work, up to certificates
