+++ sdk/keyvault/keyvault-certificates/src/lro/delete/poller.ts
@@ -1,38 +1,22 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { delay, RequestOptionsBase } from "@azure/core-http";
-import { Poller } from "@azure/core-lro";
-import {
-  DeleteCertificatePollOperationState,
-  makeDeleteCertificatePollOperation
-} from "./operation";
-import { DeletedCertificate, CertificateClientInterface } from "../../certificatesModels";
+import { DeleteCertificatePollOperation, DeleteCertificatePollOperationState } from "./operation";
+import { DeletedCertificate } from "../../certificatesModels";
+import { KeyVaultCertificatePoller, KeyVaultCertificatePollerOptions } from "../keyVaultCertificatePoller";
 
-export interface DeleteCertificatePollerOptions {
-  client: CertificateClientInterface;
-  certificateName: string;
-  requestOptions?: RequestOptionsBase;
-  intervalInMs?: number;
-  resumeFrom?: string;
-}
+export interface DeleteCertificatePollerOptions extends KeyVaultCertificatePollerOptions { }
 
 /**
  * Class that deletes a poller that waits until a certificate finishes being deleted
  * @internal
  */
-export class DeleteCertificatePoller extends Poller<
+export class DeleteCertificatePoller extends KeyVaultCertificatePoller<
   DeleteCertificatePollOperationState,
   DeletedCertificate
-> {
-  /**
-   * Defines how much time the poller is going to wait before making a new request to the service.
-   * @memberof DeleteCertificatePoller
-   */
-  public intervalInMs: number;
-
+  > {
   constructor(options: DeleteCertificatePollerOptions) {
-    const { client, certificateName, requestOptions, intervalInMs = 2000, resumeFrom } = options;
+    const { vaultUrl, client, certificateName, requestOptions, intervalInMs = 2000, resumeFrom } = options;
 
     let state: DeleteCertificatePollOperationState | undefined;
 
@@ -40,23 +24,18 @@ export class DeleteCertificatePoller extends Poller<
       state = JSON.parse(resumeFrom).state;
     }
 
-    const operation = makeDeleteCertificatePollOperation({
-      ...state,
-      certificateName,
+    const operation = new DeleteCertificatePollOperation(
+      {
+        ...state,
+        certificateName,
+      },
+      vaultUrl,
+      client,
       requestOptions,
-      client
-    });
+    );
 
     super(operation);
 
     this.intervalInMs = intervalInMs;
-  }
-
-  /**
-   * The method used by the poller to wait before attempting to update its operation.
-   * @memberof DeleteCertificatePoller
-   */
-  async delay(): Promise<void> {
-    return delay(this.intervalInMs);
   }
 }
