+++ sdk/keyvault/keyvault-certificates/src/lro/recover/operation.ts
@@ -2,131 +2,144 @@
 // Licensed under the MIT license.
 
 import { AbortSignalLike } from "@azure/abort-controller";
-import { PollOperationState, PollOperation } from "@azure/core-lro";
-import { RequestOptionsBase } from "@azure/core-http";
+import { operationOptionsToRequestOptionsBase, RequestOptionsBase } from "@azure/core-http";
+import { createSpan, setParentSpan } from "../../../../keyvault-common";
 import {
-  KeyVaultCertificateWithPolicy,
-  CertificateClientInterface
+  GetCertificateOptions,
+  KeyVaultCertificateWithPolicy, RecoverDeletedCertificateOptions
 } from "../../certificatesModels";
+import { KeyVaultClient } from "../../generated/keyVaultClient";
+import { GetCertificateResponse, RecoverDeletedCertificateResponse } from "../../generated/models";
+import { getCertificateWithPolicyFromCertificateBundle } from "../../transformations";
+import { KeyVaultCertificatePollOperation, KeyVaultCertificatePollOperationState } from "../keyVaultCertificatePoller";
 
 /**
- * The public representation of the RecoverDeletedCertificatePoller operation state.
+ * Deprecated: Public representation of the recover delete certificate poll operation
  */
-export type RecoverDeletedCertificateState = PollOperationState<KeyVaultCertificateWithPolicy>;
+export interface RecoverDeletedCertificateState
+  extends KeyVaultCertificatePollOperationState<KeyVaultCertificateWithPolicy> {
+}
 
 /**
- * An interface representing the state of a delete certificate's poll operation
- * @internal
+ * An interface representing the state of a recover delete certificate poll operation
  */
 export interface RecoverDeletedCertificatePollOperationState
-  extends PollOperationState<KeyVaultCertificateWithPolicy> {
-  /**
-   * The name of the certificate.
-   */
-  certificateName: string;
-  /**
-   * Options for the core-http requests.
-   */
-  requestOptions?: RequestOptionsBase;
-  /**
-   * An interface representing a CertificateClient. For internal use.
-   */
-  client: CertificateClientInterface;
+  extends KeyVaultCertificatePollOperationState<KeyVaultCertificateWithPolicy> {
 }
 
 /**
  * An interface representing a delete certificate's poll operation
- * @internal
  */
-export type RecoverDeletedCertificatePollOperation = PollOperation<
+export class RecoverDeletedCertificatePollOperation extends KeyVaultCertificatePollOperation<
   RecoverDeletedCertificatePollOperationState,
   KeyVaultCertificateWithPolicy
->;
-
-/**
- * @summary Reaches to the service and updates the delete certificate's poll operation.
- * @param [options] The optional parameters, which are an abortSignal from @azure/abort-controller and a function that triggers the poller's onProgress function.
- * @internal
- */
-async function update(
-  this: RecoverDeletedCertificatePollOperation,
-  options: {
-    abortSignal?: AbortSignalLike;
-    fireProgress?: (state: RecoverDeletedCertificatePollOperationState) => void;
-  } = {}
-): Promise<RecoverDeletedCertificatePollOperation> {
-  const state = this.state;
-  const { certificateName, client } = state;
-
-  const requestOptions = state.requestOptions || {};
-  if (options.abortSignal) {
-    requestOptions.abortSignal = options.abortSignal;
+  > {
+  constructor(
+    public state: RecoverDeletedCertificatePollOperationState,
+    private vaultUrl: string,
+    private client: KeyVaultClient,
+    private requestOptions: RequestOptionsBase = {}
+  ) {
+    super(state, { cancelMessage: "Canceling the deletion of a certificate is not supported." });
   }
 
-  if (!state.isStarted) {
+  /**
+* Gets the latest information available from a specific certificate, including the certificate's policy. This operation requires the certificates/get permission.
+*/
+  private async getCertificate(
+    certificateName: string,
+    options: GetCertificateOptions = {}
+  ): Promise<KeyVaultCertificateWithPolicy> {
+    const requestOptions = operationOptionsToRequestOptionsBase(options);
+    const span = createSpan("getCertificate", requestOptions);
+
+    let result: GetCertificateResponse;
+
     try {
-      state.result = await client.getCertificate(certificateName, { requestOptions });
-      state.isCompleted = true;
-    } catch (e) {
-      // getCertificate will only work once the LRO is completed.
+      result = await this.client.getCertificate(
+        this.vaultUrl,
+        certificateName,
+        "",
+        setParentSpan(span, requestOptions)
+      );
+    } finally {
+      span.end();
     }
-    if (!state.isCompleted) {
-      state.result = await client.recoverDeletedCertificate(certificateName, { requestOptions });
-      state.isStarted = true;
-    }
+
+    return getCertificateWithPolicyFromCertificateBundle(result);
   }
 
-  if (!state.isCompleted) {
+  /**
+   * Recovers the deleted certificate in the specified vault. This operation can only be performed on a soft-delete enabled vault. This operation
+   * requires the certificate/recover permission.
+   */
+  private async recoverDeletedCertificate(
+    certificateName: string,
+    options: RecoverDeletedCertificateOptions = {}
+  ): Promise<KeyVaultCertificateWithPolicy> {
+    const requestOptions = operationOptionsToRequestOptionsBase(options);
+    const span = createSpan("recoverDeletedCertificate", requestOptions);
+
+    let result: RecoverDeletedCertificateResponse;
+
     try {
-      state.result = await client.getCertificate(certificateName, { requestOptions });
-      state.isCompleted = true;
-    } catch (error) {
-      if (error.statusCode === 403) {
-        // At this point, the resource exists but the user doesn't have access to it.
+      result = await this.client.recoverDeletedCertificate(
+        this.vaultUrl,
+        certificateName,
+        setParentSpan(span, requestOptions)
+      );
+    } finally {
+      span.end();
+    }
+
+    return getCertificateWithPolicyFromCertificateBundle(result._response.parsedBody);
+  }
+
+
+  /**
+   * Reaches to the service and updates the delete certificate's poll operation.
+   */
+  async update(
+    options: {
+      abortSignal?: AbortSignalLike;
+      fireProgress?: (state: RecoverDeletedCertificatePollOperationState) => void;
+    } = {}
+  ): Promise<RecoverDeletedCertificatePollOperation> {
+    const state = this.state;
+    const { certificateName } = state;
+
+    if (options.abortSignal) {
+      this.requestOptions.abortSignal = options.abortSignal;
+    }
+
+    if (!state.isStarted) {
+      try {
+        state.result = await this.getCertificate(certificateName, this.requestOptions);
         state.isCompleted = true;
-      } else if (error.statusCode !== 404) {
-        state.error = error;
-        state.isCompleted = true;
+      } catch (e) {
+        // getCertificate will only work once the LRO is completed.
+      }
+      if (!state.isCompleted) {
+        state.result = await this.recoverDeletedCertificate(certificateName, this.requestOptions);
+        state.isStarted = true;
       }
     }
+
+    if (!state.isCompleted) {
+      try {
+        state.result = await this.getCertificate(certificateName, this.requestOptions);
+        state.isCompleted = true;
+      } catch (error) {
+        if (error.statusCode === 403) {
+          // At this point, the resource exists but the user doesn't have access to it.
+          state.isCompleted = true;
+        } else if (error.statusCode !== 404) {
+          state.error = error;
+          state.isCompleted = true;
+        }
+      }
+    }
+
+    return this;
   }
-
-  return makeRecoverDeletedCertificatePollOperation(state);
-}
-
-/**
- * @summary Reaches to the service and cancels the certificate's operation, also updating the certificate's poll operation
- * @param [options] The optional parameters, which is only an abortSignal from @azure/abort-controller
- * @internal
- */
-async function cancel(): Promise<RecoverDeletedCertificatePollOperation> {
-  throw new Error("Canceling the deletion of a certificate is not supported.");
-}
-
-/**
- * @summary Serializes the create certificate's poll operation
- * @internal
- */
-function toString(this: RecoverDeletedCertificatePollOperation): string {
-  return JSON.stringify({
-    state: this.state
-  });
-}
-
-/**
- * @summary Builds a create certificate's poll operation
- * @param [state] A poll operation's state, in case the new one is intended to follow up where the previous one was left.
- * @internal
- */
-export function makeRecoverDeletedCertificatePollOperation(
-  state: RecoverDeletedCertificatePollOperationState
-): RecoverDeletedCertificatePollOperation {
-  return {
-    state: {
-      ...state
-    },
-    update,
-    cancel,
-    toString
-  };
 }
