+++ sdk/keyvault/keyvault-common/src/index.ts
@@ -3,3 +3,4 @@
 
 export * from "./challengeBasedAuthenticationPolicy";
 export * from "./parseKeyvaultIdentifier";
+export * from "./tracing";
