+++ sdk/keyvault/keyvault-secrets/src/lro/delete/operation.ts
@@ -10,14 +10,14 @@ import {
 } from "../keyVaultSecretPoller";
 import { KeyVaultClient } from "../../generated/keyVaultClient";
 import { DeleteSecretResponse, GetDeletedSecretResponse } from "../../generated/models";
-import { createSpan, setParentSpan } from "../../tracing";
+import { createSpan, setParentSpan } from "../../../../keyvault-common/src";
 import { getSecretFromSecretBundle } from "../../transformations";
 
 /**
  * An interface representing the state of a delete secret's poll operation
  */
 export interface DeleteSecretPollOperationState
-  extends KeyVaultSecretPollOperationState<DeletedSecret> {}
+  extends KeyVaultSecretPollOperationState<DeletedSecret> { }
 
 /**
  * An interface representing a delete secret's poll operation
@@ -25,7 +25,7 @@ export interface DeleteSecretPollOperationState
 export class DeleteSecretPollOperation extends KeyVaultSecretPollOperation<
   DeleteSecretPollOperationState,
   DeletedSecret
-> {
+  > {
   constructor(
     public state: DeleteSecretPollOperationState,
     private vaultUrl: string,
@@ -38,8 +38,6 @@ export class DeleteSecretPollOperation extends KeyVaultSecretPollOperation<
   /**
    * Sends a delete request for the given Key Vault Key's name to the Key Vault service.
    * Since the Key Vault Key won't be immediately deleted, we have {@link beginDeleteKey}.
-   * @param {string} name The name of the Key Vault Key.
-   * @param {DeleteKeyOptions} [options] Optional parameters for the underlying HTTP request.
    */
   private async deleteSecret(
     name: string,
@@ -65,9 +63,6 @@ export class DeleteSecretPollOperation extends KeyVaultSecretPollOperation<
   /**
    * The getDeletedSecret method returns the specified deleted secret along with its properties.
    * This operation requires the secrets/get permission.
-   * @summary Gets the specified deleted secret.
-   * @param {string} name The name of the secret.
-   * @param {GetDeletedKeyOptions} [options] The optional parameters.
    */
   private async getDeletedSecret(
     name: string,
@@ -91,8 +86,7 @@ export class DeleteSecretPollOperation extends KeyVaultSecretPollOperation<
   }
 
   /**
-   * @summary Reaches to the service and updates the delete secret's poll operation.
-   * @param [options] The optional parameters, which are an abortSignal from @azure/abort-controller and a function that triggers the poller's onProgress function.
+   * Reaches to the service and updates the delete secret's poll operation.
    */
   public async update(
     options: {
